justify queue interfaces
