+++ spec/models/slide_spec.rb
@@ -25,7 +25,7 @@ require 'rails_helper'
 
 describe 'Slide' do
   before do
-    SlideHub::Cloud::AWSEngine.configure do |config|
+    SlideHub::Cloud::Engine::AWS.configure do |config|
       config.region = 'ap-northeast-1'
       config.aws_access_id = 'aws_access_id'
       config.aws_secret_key = 'aws_secret_key'
